Split non-diff stuff to new module
